pavouci_api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
sqladmin
